refactor: landing페이지 컨텐츠 통합
