+++ src/pages/Landing/index.jsx
@@ -1,7 +1,10 @@
 import Banner from "./Banner";
 import Feature from "./Feature";
-import { heroBannerData, footerBannerData } from "./BannerContent";
-import { featureList } from "./featureContent";
+import {
+  heroBannerData,
+  footerBannerData,
+  featureList,
+} from "./landingContents";
 
 export default function Landing() {
   return (
